Update test workflow actions and filterblock location (#37)
* Update test workflow actions and filterblock location

This should get the tests workflow passing again, or at least closer
to that. The tests pass locally for me now.

Signed-off-by: Ben Browning <bbrownin@redhat.com>

* Disable e2e workflow with a note about why

We don't have (nor want to use) `ubuntu-gpu` runners in this GitHub
org. So, setting the e2e workflow to be manually dispatched until we
adjust it to use AWS GPU runners that get provisioned during the
workflow and until we adjust the e2e test to only test SDG parts
instead of the entire InstructLab e2e tests.

Signed-off-by: Ben Browning <bbrownin@redhat.com>

---------

Signed-off-by: Ben Browning <bbrownin@redhat.com>
Co-authored-by: Kai Xu <me@xuk.ai>